Database connection for Seminar.
Data in seminar table comes now from backend.
I have also made connection for dozent and student. We need the
related student and dozent for a seminar. Thats why we have methods that
brings a single student and dozent with a given id.

Also fontwaseome implemented to the projekt. We are using it to use
icons.